Separate molpro_input.py
